+++ _config/types/config.footer.types.ts
