+++ handler/actionHandler/index.ts
@@ -11,7 +11,7 @@ import {
   IUpdateAction,
   IParallelAction,
   IConditionAction,
-} from '_config/config.handler';
+} from '_config/types/config.handler.types';
 import insertActionHandler from './insert';
 import mutateActionHandler from './mutate';
 import selectActionHandler from './select';
