+++ paths.ts
@@ -10,27 +10,27 @@ export const ENTITY_MAP_PATH = mergePaths(ROOT_PATH, '/generated/entityMap.ts');
 
 export const PAGE_CONFIG_PATH = mergePaths(
   ROOT_PATH,
-  '/_config/config.page.ts',
+  '/_config/types/config.page.types.ts',
 );
 
 export const MENU_CONFIG_PATH = mergePaths(
   ROOT_PATH,
-  '/_config/config.menu.ts',
+  '/src/_config/types/config.menu.types',
 );
 
 export const MODAL_CONFIG_PATH = mergePaths(
   ROOT_PATH,
-  '/_config/config.modal.ts',
+  '/src/_config/types/config.modal.types',
 );
 
 export const SIDEPANEL_CONFIG_PATH = mergePaths(
   ROOT_PATH,
-  '/_config/config.sidepanel.ts',
+  '/src/_config/types/config.sidepanel.types',
 );
 
 export const FOOTER_CONFIG_PATH = mergePaths(
   ROOT_PATH,
-  '/_config/config.footer.ts',
+  '/src/_config/types/config.footer.types',
 );
 
 export const SERVER_ROOT_PATH = mergePaths(process.cwd(), '../platform-server');
@@ -52,25 +52,25 @@ export const SERVER_ENTITY_MAP_PATH = mergePaths(
 
 export const SERVER_PAGE_CONFIG_PATH = mergePaths(
   SERVER_ROOT_PATH,
-  '/src/config/config.page.ts',
+  '/src/_config/types/config.page.types.ts',
 );
 
 export const SERVER_MENU_CONFIG_PATH = mergePaths(
   SERVER_ROOT_PATH,
-  '/src/config/config.menu.ts',
+  '/src/_config/types/config.menu.types',
 );
 
 export const SERVER_MODAL_CONFIG_PATH = mergePaths(
   SERVER_ROOT_PATH,
-  '/src/config/config.modal.ts',
+  '/src/_config/types/config.modal.types',
 );
 
 export const SERVER_SIDEPANEL_CONFIG_PATH = mergePaths(
   SERVER_ROOT_PATH,
-  '/src/config/config.sidepanel.ts',
+  '/src/_config/types/config.sidepanel.types',
 );
 
 export const SERVER_FOOTER_CONFIG_PATH = mergePaths(
   SERVER_ROOT_PATH,
-  '/src/config/config.footer.ts',
+  '/src/_config/types/config.footer.types',
 );
